Add GMGame and PlayerGame pages
