+++ arpui/Cargo.toml
@@ -20,6 +20,7 @@ log = "0.4.19"
 reqwest = { version = "0.12.3", features = ["json"] }
 wasm-bindgen = "0.2.92"
 wasm-cookies = "0.2.1"
+uuid = { version = "1.1.2", features = ["v4", "serde"] }
 
 [dependencies.web-sys]
 version = "0.3.69"
