+++ arpui/src/main.rs
@@ -2,23 +2,24 @@
 
 use dioxus::prelude::*;
 use log::{info, LevelFilter};
+use uuid::Uuid;
 
-use arptypes::types::Dice;
-use arptypes::multitenant;
+mod rpi;
+use rpi::{auth_token, AUTH_TOKEN, list_games};
 
 #[derive(Clone, Routable, Debug, PartialEq)]
 enum Route {
   #[route("/")]
   Main,
-  #[route("/test")]
-  Test,
-  #[route("/blog/:id")]
-  Blog { id: i32 },
+  #[route("/gm/:id")]
+  GMGame { id: Uuid },
+  #[route("/player/:id")]
+  PlayerGame { id: Uuid },
 }
 
 fn main() {
   // Init debug
-  dioxus_logger::init(LevelFilter::Info).expect("failed to init logger");
+  dioxus_logger::init(LevelFilter::Trace).expect("failed to init logger");
   console_error_panic_hook::set_once();
 
   launch(App);
@@ -28,19 +29,6 @@ fn App() -> Element {
   rsx! { Router::<Route> {} }
 }
 
-#[component]
-fn Blog(id: i32) -> Element {
-  rsx! {
-    Link { to: Route::Main, "Go Home" }
-    "Blog post {id}"
-  }
-}
-
-static AUTH_TOKEN: GlobalSignal<String> = Signal::global(String::new);
-
-fn auth_token() -> String {
-  wasm_cookies::get("arpeggio-token").unwrap_or(Ok(String::new())).unwrap_or(String::new())
-}
 
 #[component]
 fn Main() -> Element {
@@ -51,7 +39,6 @@ fn Main() -> Element {
   }
 
   rsx! {
-    div { Link { to: Route::Test, "Test stuff" } }
     div { "Oh no, the cookie {AUTH_TOKEN():?} "}
     div {
       "Enter your auth token:"
@@ -67,20 +54,6 @@ fn Main() -> Element {
   }
 }
 
-fn rpi_url() -> String {
-  let window = web_sys::window().expect("global window doesn't exist");
-  let document = window.document().expect("document must exist");
-  let meta = document.query_selector("meta[name='RPI_URL']").expect("meta RPI_URL tag must exist in index.html (1)").expect("meta RPI_URL tag must exist in index.html (2)");
-  meta.get_attribute("content").expect("meta RPI_URL tag must have content")
-}
-
-async fn list_games() -> Result<multitenant::GameList, reqwest::Error> {
-  let rpi_url = rpi_url();
-  let url = format!("{rpi_url}/g/list");
-  let client = reqwest::Client::new();
-  let response = client.get(url).header("x-arpeggio-auth", AUTH_TOKEN()).send().await?;
-  response.json().await
-}
 
 #[component]
 fn GameList() -> Element {
@@ -91,44 +64,26 @@ fn GameList() -> Element {
       "got a list!"
       ul {
         for (profile, metadata) in &list.games {
-          li { "{profile.profile_name} @ {metadata.name} (as {profile.role})" }
+          li {
+            a {
+              href: "/{profile.role.to_string().to_lowercase()}/{profile.game_id}",
+              "{profile.profile_name} @ {metadata.name} (as {profile.role})" } }
         }
       }
     },
     Some(Err(e)) => rsx! { "Error! {e:?}" },
-    None => rsx! {"Promise not fulfilled"}
+    None => rsx! {"Loading games!"}
   }
 
 }
 
+
 #[component]
-fn Test() -> Element {
-  let mut count = use_signal(|| 0);
+fn GMGame(id: Uuid) -> Element {
+  rsx! { "id: {id:?}"}
+}
 
-  let dice = Dice::Expr { num: 3, size: 6 };
-
-  rsx! {
-    ul {
-      li { Link { to: Route::Blog { id: count() }, "Go to blog" } }
-      li { Link { to: Route::Main, "Go Home" } }
-    }
-    div {
-      h1 { "High-Five counter: {count}" }
-      div { "Dice expression: {dice:?}" }
-      button {
-        onclick: move |_| {
-            count += 1;
-            info!("UP {count}");
-        },
-        "Up high!"
-      }
-      button {
-        onclick: move |_| {
-            count -= 1;
-            info!("DOWN {count}");
-        },
-        "Down low!"
-      }
-    }
-  }
+#[component]
+fn PlayerGame(id: Uuid) -> Element {
+  rsx! { "id: {id:?}"}
 }
